feat: add Supabase SSR clients and session middleware
